requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.9.3
bilibili-api-python==11.8.1
bilibili-storytree==0.1.1
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
fonttools==4.34.4
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
mgp-common==0.1.7
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
Pillow==9.1.1
pycryptodome==3.15.0
PyExecJS==1.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
pywikibot==7.5.0
PyYAML==5.4.1
qrcode==7.3.1
regex==2022.7.25
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
urllib3==1.26.11
wcwidth==0.2.5
wikitextparser==0.49.4
yarl==1.8.0
